DRY - used same component to both tabs
