+++ app.json
@@ -3,7 +3,7 @@
     "name": "IsracardDemoNaorBarKochva",
     "slug": "IsracardDemoNaorBarKochva",
     "version": "1.0.0",
-    "orientation": "portrait",
+    "orientation": "default",
     "icon": "./assets/icon.png",
     "userInterfaceStyle": "light",
     "newArchEnabled": true,
